storyboard_git/requirements.txt
websockets
pymupdf
python-pptx
crewai
requests
python-dotenv
asyncio